Refactoring the command management... sigh that took a while
