--- plusplusbot/commands/general_commands/__init__.py
@@ -1,6 +1,0 @@
-from .help_command import HelpCommand
-
-
-general_commands_list = [
-    HelpCommand,
-]
