+++ plusplusbot/handlers.py
@@ -1,4 +1,3 @@
-
 import botocore
 import pathlib
 import logging
