+++ plusplusbot/helpers.py
@@ -1,5 +1,5 @@
-import re
 import string
+import re
 
 from unidecode import unidecode
